+++ front/src/components/theme/mode-toggle.tsx
@@ -1,6 +1,6 @@
 import { Moon, Sun } from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { useTheme } from "@/components/theme-provider";
+import { useTheme } from "@/components/theme/theme-provider";
 
 export function ModeToggle() {
   const { theme, setTheme } = useTheme();
